Mcleod Updates
I've added a new layout for the api results using D3. I played around with how to layout the data visually, when i don' t know how many results will be returned. I settled on using the force layout which will allow me to pack the data in neatly regardless of number. I'd also like to replace the circles i'm using as placeholders with some other sort of visual, either the labels or the glass images. the visual and interaction would then be based on the d3 layout vs the tables. The food results i'm hoping can be fed in as a pack layout (circle within circle). If i can massage the data structure enough. The main issue now that design decisions have been made is restructuring my code to be clearer and more efficient. It would also be nice if i could call you tube for how to make a recipe once i have the food results.